Add installer project and icons, update license
Introduced a new Installer project using WiX, including all required assets, UI dialogs, and configuration files. Added application and installer icons to the Editor project and set the application icon in the project file. Updated the license format and fixed a typo in the data size converter. Updated CI workflow to build the installer and added the Installer project to the solution.